refactor(direct-control): extract ready city source
Move the ready-city embedded App UI source into src/play/ready/city.ts while keeping the public wrapper and command builder in the facade. Keep the ready-city production, town-focus, and population-placement helpers inside that owner for this slice rather than creating a broad shared ready catalog.

Record the ready atom set as extracted in the support modularization OpenSpec workstream. This is source ownership proof only and does not claim live-game runtime proof.

Validation: bun run --cwd packages/civ7-direct-control test -- ready-city-view.test.ts; bun run --cwd packages/civ7-direct-control test; bun run --cwd packages/civ7-direct-control check; bun run --cwd packages/civ7-direct-control build; bun run openspec -- validate civ7-support-direct-control-modularization --strict; bun run resources:status; git diff --check.
