+++ openspec/changes/civ7-support-direct-control-modularization/tasks.md
@@ -114,10 +114,12 @@ runtime/direct-control claims.
     readback edge coverage to owning package suites before source extraction.
 - [ ] 4.3 Extract notification view/materialization atom.
 - [ ] 4.4 Extract notification dismissal/verification atom.
-- [ ] 4.5 Extract ready unit/city view atoms.
+- [x] 4.5 Extract ready unit/city view atoms.
   - [x] 4.5.1 Extract unit move preview embedded source owner while keeping
         the public wrapper in the facade.
   - [x] 4.5.2 Extract ready-unit embedded source owner while keeping the public
+        wrapper in the facade.
+  - [x] 4.5.3 Extract ready-city embedded source owner while keeping the public
         wrapper in the facade.
 - [ ] 4.6 Extract operation validation/send/postcondition atoms.
 - [ ] 4.7 Extract settlement/tactical/progression read atoms.
